Pause the Iterator thread when there are too many un-sent streaming rows.
We keep track of the number of iterated rows and the number of rows dart has added to the
stream controller. The difference between these numbers is the number of rows in the IPC buffer.

If this number gets too big then pause the worker thread. It will be woken up again when all
buffered rows have been processed.